Change to separate files and use css vars over js object
