+++ app/components/shared.js
@@ -1,0 +1,21 @@
+const baseStyles = `
+  * {
+    --color-primary: lavender;
+    --color-highlight: deeppink;
+    --color-magnolia: #f7f7f7;
+    --color-dark: #0d0d0d;
+
+    box-sizing: border-box;
+    margin: 0;
+    color: var(--color-dark);
+  }
+
+  button {
+    background: none;
+    border: none;
+    outline: none;
+    padding: 0;
+  }
+`;
+
+export { baseStyles };
